+++ clisops/ops/average.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 from typing import List, Optional, Sequence, Tuple, Union
 
+import geopandas as gpd
 import xarray as xr
 from roocs_utils.exceptions import InvalidParameterValue
 from roocs_utils.parameter.dimension_parameter import DimensionParameter
@@ -10,7 +11,7 @@ from clisops.core import average
 from clisops.ops.base_operation import Operation
 from clisops.utils.file_namers import get_file_namer
 
-__all__ = ["average_over_dims", "average_time"]
+__all__ = ["average_over_dims", "average_time", "average_shape"]
 
 
 class Average(Operation):
@@ -70,8 +71,8 @@ def average_over_dims(
     Returns
     -------
     List[Union[xr.Dataset, str]]
-    A list of the outputs in the format selected; str corresponds to file paths if the
-    output format selected is a file.
+        A list of the outputs in the format selected; str corresponds to file paths if the
+        output format selected is a file.
 
     Examples
     --------
@@ -85,6 +86,81 @@ def average_over_dims(
 
     """
     op = Average(**locals())
+    return op.process()
+
+
+class AverageShape(Operation):
+    def _resolve_params(self, **params):
+        shape = params.get("shape")
+        variable = params.get("variable", None)
+
+        self.params = {"shape": shape, "variable": variable}
+
+        if not shape:
+            raise InvalidParameterValue(
+                "At least one area for averaging must be provided"
+            )
+
+    def _get_file_namer(self):
+        extra = "_avg-shape"
+
+        namer = get_file_namer(self._file_namer)(extra=extra)
+
+        return namer
+
+    def _calculate(self):
+        avg_ds = average.average_shape(
+            self.ds,
+            self.params.get("shape", None),
+            self.params.get("variable", None),
+        )
+        return avg_ds
+
+
+def average_shape(
+    ds: Union[xr.Dataset, Path, str],
+    shape: Union[str, Path, gpd.GeoDataFrame],
+    variable: Optional[Union[str, Sequence[str]]] = None,
+    output_dir: Optional[Union[str, Path]] = None,
+    output_type: str = "netcdf",
+    split_method: str = "time:auto",
+    file_namer: str = "standard",
+) -> List[Union[xr.Dataset, str]]:
+    """
+
+    Parameters
+    ----------
+    ds : Union[xr.Dataset, str]
+        Xarray dataset.
+    shape : Union[str, Path, gpd.GeoDataFrame]
+        Path to shape file, or directly a GeoDataFrame. Supports formats compatible with geopandas.
+        Will be converted to EPSG:4326 if needed.
+
+    variable : Optional[Union[str, Sequence[str], None]]
+        Variables to average. If None, average over all data variables.
+    output_dir : Optional[Union[str, Path]]
+    output_type : {"netcdf", "nc", "zarr", "xarray"}
+    split_method : {"time:auto"}
+    file_namer : {"standard", "simple"}
+
+    Returns
+    -------
+    List[Union[xr.Dataset, str]]
+        A list of the outputs in the format selected; str corresponds to file paths if the
+        output format selected is a file.
+
+    Examples
+    --------
+    | ds: xarray Dataset or "cmip5.output1.MOHC.HadGEM2-ES.rcp85.mon.atmos.Amon.r1i1p1.latest.tas"
+    | dims: ['latitude', 'longitude']
+    | ignore_undetected_dims: False
+    | output_dir: "/cache/wps/procs/req0111"
+    | output_type: "netcdf"
+    | split_method: "time:auto"
+    | file_namer: "standard"
+
+    """
+    op = AverageShape(**locals())
     return op.process()